backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
supabase==0.5.8
python-dotenv==0.19.0
pydantic==1.8.2